requirements.txt
alembic==1.0.10
Babel==2.7.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.11
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.4
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
